Move all project wizard stuff into lambda package, standardize language specific parts location
