+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/java/JavaSamProjectWizard.kt
@@ -21,18 +21,18 @@ import software.amazon.awssdk.services.lambda.model.Runtime
 import software.aws.toolkits.core.utils.getLogger
 import software.aws.toolkits.core.utils.logWhenNull
 import software.aws.toolkits.jetbrains.services.lambda.BuiltInRuntimeGroups
-import software.aws.toolkits.jetbrains.services.lambda.SamNewProjectSettings
-import software.aws.toolkits.jetbrains.services.lambda.SamProjectTemplate
-import software.aws.toolkits.jetbrains.services.lambda.SamProjectWizard
-import software.aws.toolkits.jetbrains.services.lambda.TemplateParameters
-import software.aws.toolkits.jetbrains.services.lambda.TemplateParameters.AppBasedTemplate
 import software.aws.toolkits.jetbrains.services.lambda.sam.SamSchemaDownloadPostCreationAction
+import software.aws.toolkits.jetbrains.services.lambda.wizard.IntelliJSdkSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamNewProjectSettings
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectGenerator
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectTemplate
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectWizard
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SchemaResourceSelectorSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SchemaSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SdkSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.TemplateParameters
+import software.aws.toolkits.jetbrains.services.lambda.wizard.TemplateParameters.AppBasedTemplate
 import software.aws.toolkits.jetbrains.services.schemas.SchemaCodeLangs
-import software.aws.toolkits.jetbrains.ui.wizard.IntelliJSdkSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SamProjectGenerator
-import software.aws.toolkits.jetbrains.ui.wizard.SchemaResourceSelectorSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SchemaSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SdkSelectionPanel
 import software.aws.toolkits.resources.message
 import java.nio.file.Paths
