+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/python/PythonSamProjectWizard.kt
@@ -10,21 +10,20 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.util.PlatformUtils
 import software.amazon.awssdk.services.lambda.model.Runtime
 import software.aws.toolkits.jetbrains.services.lambda.BuiltInRuntimeGroups
-import software.aws.toolkits.jetbrains.services.lambda.SamNewProjectSettings
-import software.aws.toolkits.jetbrains.services.lambda.SamProjectTemplate
-import software.aws.toolkits.jetbrains.services.lambda.SamProjectWizard
-import software.aws.toolkits.jetbrains.services.lambda.TemplateParameters
-import software.aws.toolkits.jetbrains.services.lambda.TemplateParameters.AppBasedTemplate
-import software.aws.toolkits.jetbrains.services.lambda.TemplateParameters.LocationBasedTemplate
 import software.aws.toolkits.jetbrains.services.lambda.sam.SamCommon
 import software.aws.toolkits.jetbrains.services.lambda.sam.SamSchemaDownloadPostCreationAction
+import software.aws.toolkits.jetbrains.services.lambda.wizard.IntelliJSdkSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamNewProjectSettings
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectGenerator
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectTemplate
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SamProjectWizard
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SchemaResourceSelectorSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SchemaSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.SdkSelectionPanel
+import software.aws.toolkits.jetbrains.services.lambda.wizard.TemplateParameters
+import software.aws.toolkits.jetbrains.services.lambda.wizard.TemplateParameters.AppBasedTemplate
+import software.aws.toolkits.jetbrains.services.lambda.wizard.TemplateParameters.LocationBasedTemplate
 import software.aws.toolkits.jetbrains.services.schemas.SchemaCodeLangs
-import software.aws.toolkits.jetbrains.ui.wizard.IntelliJSdkSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.PyCharmSdkSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SamProjectGenerator
-import software.aws.toolkits.jetbrains.ui.wizard.SchemaResourceSelectorSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SchemaSelectionPanel
-import software.aws.toolkits.jetbrains.ui.wizard.SdkSelectionPanel
 import software.aws.toolkits.resources.message
 import java.nio.file.Paths
