+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/lambda/wizard/SchemaCodeGenUtilsTest.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.ui.wizard
+package software.aws.toolkits.jetbrains.services.lambda.wizard
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
